requirements.txt
# Azure Agent package (exact version)
azure-ai-projects==2.0.0b3
azure-storage-blob
azure-cosmos
six

# Your top-level deps that you intentionally want
fastapi
uvicorn
openai
anthropic

python-dotenv
requests
pydantic
pydantic-settings

chainlit
agent-framework-core
agent-framework-azure-ai

asyncpg 
SQLAlchemy 
azure-identity 
azure-storage-file-datalake 
aiohttp 
greenlet
psycopg2